+++ packages/crdt_lf/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [Unreleased]
+
+### Added
+- Tests
+  
+### Fixed
+- Fix cached value in handlers
+
 ## [0.1.0] - 2025-04-01
 
 Initial release